+++ src/lib.rs
@@ -42,7 +42,6 @@ pub mod fetch_stage;
 pub mod fullnode;
 pub mod gossip_service;
 pub mod leader_scheduler;
-pub mod ledger;
 pub mod mint;
 pub mod packet;
 pub mod poh;
