+++ src/thin_client.rs
@@ -427,9 +427,9 @@ mod tests {
     use super::*;
     use crate::bank::Bank;
     use crate::cluster_info::Node;
+    use crate::db_ledger::create_tmp_ledger_with_mint;
     use crate::fullnode::Fullnode;
     use crate::leader_scheduler::LeaderScheduler;
-    use crate::ledger::create_tmp_ledger_with_mint;
 
     use crate::mint::Mint;
     use bincode::deserialize;
