+++ src/tvu.rs
@@ -173,11 +173,11 @@ impl Service for Tvu {
 pub mod tests {
     use crate::bank::Bank;
     use crate::cluster_info::{ClusterInfo, Node};
+    use crate::db_ledger::get_tmp_ledger_path;
     use crate::db_ledger::DbLedger;
     use crate::entry::Entry;
     use crate::gossip_service::GossipService;
     use crate::leader_scheduler::LeaderScheduler;
-    use crate::ledger::get_tmp_ledger_path;
 
     use crate::mint::Mint;
     use crate::packet::SharedBlob;
