+++ tests/replicator.rs
@@ -9,10 +9,10 @@ use solana::client::mk_client;
 use solana::cluster_info::{ClusterInfo, Node, NodeInfo};
 use solana::create_vote_account::*;
 use solana::db_ledger::DbLedger;
+use solana::db_ledger::{create_tmp_genesis, get_tmp_ledger_path, tmp_copy_ledger};
 use solana::entry::Entry;
 use solana::fullnode::Fullnode;
 use solana::leader_scheduler::LeaderScheduler;
-use solana::ledger::{create_tmp_genesis, get_tmp_ledger_path, tmp_copy_ledger};
 use solana::replicator::Replicator;
 use solana::rpc_request::{RpcClient, RpcRequest};
 use solana::streamer::blob_receiver;
